Moved Logger, VerifyJWT, and GetJwtMessageVerify function out of the router.
