+++ NodeJsFiles/utils/Logger.js
@@ -1,0 +1,26 @@
+require('dotenv').config(); // Loading .env to process.env
+const winston = require('winston');
+
+/** Setting up the Winston logger.
+  * Under the development mode log to console.
+*/
+const logger = new winston.Logger({
+  level: process.env.LOGGING_LEVEL,
+  transports: [
+    new (winston.transports.Console)()
+  ]
+});
+
+/** Replaces the previous transports with those in the
+new configuration wholesale.
+  * When under the production mode, log to a file.
+*/
+if (process.env.NODE_ENV === 'production')
+  logger.configure({
+    level: 'error',
+    transports: [
+      new (winston.transports.File)({ filename: 'error.log' })
+    ]
+  });
+
+module.exports = logger;
